+++ src/AvroConvert/Features/AvroToJson/Resolver.cs
@@ -21,8 +21,10 @@
 #endregion
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using SolTechnology.Avro.AvroObjectServices.Read;
 using SolTechnology.Avro.AvroObjectServices.Schemas;
 using SolTechnology.Avro.AvroObjectServices.Schemas.Abstract;
@@ -152,6 +154,11 @@ namespace SolTechnology.Avro.Features.AvroToJson
                 readerSchema = ((ArraySchema)readerSchema).ItemSchema;
             }
 
+            if (readerSchema.Name == "KeyValuePair2")
+            {
+                return ResolveDictionary((RecordSchema)readerSchema, d);
+            }
+
             object[] result = Array.Empty<object>();
             int i = 0;
 
@@ -169,6 +176,22 @@ namespace SolTechnology.Avro.Features.AvroToJson
             }
 
             return result;
+        }
+
+        protected object ResolveDictionary(RecordSchema readerSchema, IReader d)
+        {
+            dynamic resultDictionary = new Dictionary<object, object>();
+
+            for (int n = (int)d.ReadArrayStart(); n != 0; n = (int)d.ReadArrayNext())
+            {
+                for (int j = 0; j < n; j++)
+                {
+                    dynamic key = Resolve(readerSchema.GetField("Key").TypeSchema, d);
+                    dynamic value = Resolve(readerSchema.GetField("Value").TypeSchema, d);
+                    resultDictionary.Add(key, value);
+                }
+            }
+            return resultDictionary;
         }
 
         protected virtual object ResolveUnion(UnionSchema readerSchema, IReader d)